refactor(channels): migrate internal callers onto the message-access ingress engine
