+++ extensions/qqbot/src/engine/group/message-gating.ts
@@ -1,4 +1,6 @@
 // Qqbot plugin module implements message gating behavior.
+import type { MentionGatePort } from "../adapter/mention-gate.port.js";
+
 type GroupMessageGateAction =
   | "drop_other_mention"
   | "block_unauthorized_command"
@@ -12,6 +14,7 @@ export interface GroupMessageGateResult {
 }
 
 export interface GroupMessageGateInput {
+  mentionGatePort: MentionGatePort;
   ignoreOtherMentions: boolean;
   hasAnyMention: boolean;
   wasMentioned: boolean;
@@ -20,46 +23,19 @@ export interface GroupMessageGateInput {
   isControlCommand: boolean;
   commandAuthorized: boolean;
   requireMention: boolean;
-  canDetectMention: boolean;
 }
 
-function resolveMentionGating(input: {
-  requireMention: boolean;
-  canDetectMention: boolean;
-  wasMentioned: boolean;
-  implicitMention: boolean;
-  shouldBypassMention: boolean;
-}): { effectiveWasMentioned: boolean; shouldSkip: boolean } {
-  const effectiveWasMentioned =
-    input.wasMentioned || input.implicitMention || input.shouldBypassMention;
-  const shouldSkip = input.requireMention && input.canDetectMention && !effectiveWasMentioned;
-  return { effectiveWasMentioned, shouldSkip };
-}
-
-function resolveCommandBypass(input: {
-  requireMention: boolean;
-  wasMentioned: boolean;
-  hasAnyMention: boolean;
-  allowTextCommands: boolean;
-  commandAuthorized: boolean;
-  isControlCommand: boolean;
-}): boolean {
-  return (
-    input.requireMention &&
-    !input.wasMentioned &&
-    !input.hasAnyMention &&
-    input.allowTextCommands &&
-    input.commandAuthorized &&
-    input.isControlCommand
-  );
-}
-
-export function resolveGroupMessageGate(input: GroupMessageGateInput): GroupMessageGateResult {
+/**
+ * Group gate Layer 1 (ignoreOtherMentions) is QQ-specific and decided here;
+ * Layer 2+3 (command gating + mention gating + command bypass) delegate to the
+ * mention gate port backed by the SDK's `resolveInboundMentionDecision`.
+ */
+export function resolveGroupMessageGate(params: GroupMessageGateInput): GroupMessageGateResult {
   if (
-    input.ignoreOtherMentions &&
-    input.hasAnyMention &&
-    !input.wasMentioned &&
-    !input.implicitMention
+    params.ignoreOtherMentions &&
+    params.hasAnyMention &&
+    !params.wasMentioned &&
+    !params.implicitMention
   ) {
     return {
       action: "drop_other_mention",
@@ -68,7 +44,23 @@ export function resolveGroupMessageGate(input: GroupMessageGateInput): GroupMess
     };
   }
 
-  if (input.allowTextCommands && input.isControlCommand && !input.commandAuthorized) {
+  const decision = params.mentionGatePort.resolveInboundMentionDecision({
+    facts: {
+      canDetectMention: true,
+      wasMentioned: params.wasMentioned,
+      hasAnyMention: params.hasAnyMention,
+      implicitMentionKinds: params.implicitMention ? ["reply_to_bot"] : [],
+    },
+    policy: {
+      isGroup: true,
+      requireMention: params.requireMention,
+      allowTextCommands: params.allowTextCommands,
+      hasControlCommand: params.isControlCommand,
+      commandAuthorized: params.commandAuthorized,
+    },
+  });
+
+  if (params.allowTextCommands && params.isControlCommand && !params.commandAuthorized) {
     return {
       action: "block_unauthorized_command",
       effectiveWasMentioned: false,
@@ -76,34 +68,17 @@ export function resolveGroupMessageGate(input: GroupMessageGateInput): GroupMess
     };
   }
 
-  const shouldBypassMention = resolveCommandBypass({
-    requireMention: input.requireMention,
-    wasMentioned: input.wasMentioned,
-    hasAnyMention: input.hasAnyMention,
-    allowTextCommands: input.allowTextCommands,
-    commandAuthorized: input.commandAuthorized,
-    isControlCommand: input.isControlCommand,
-  });
-
-  const mentionGate = resolveMentionGating({
-    requireMention: input.requireMention,
-    canDetectMention: input.canDetectMention,
-    wasMentioned: input.wasMentioned,
-    implicitMention: input.implicitMention,
-    shouldBypassMention,
-  });
-
-  if (mentionGate.shouldSkip) {
+  if (decision.shouldSkip) {
     return {
       action: "skip_no_mention",
-      effectiveWasMentioned: mentionGate.effectiveWasMentioned,
-      shouldBypassMention,
+      effectiveWasMentioned: decision.effectiveWasMentioned,
+      shouldBypassMention: decision.shouldBypassMention,
     };
   }
 
   return {
     action: "pass",
-    effectiveWasMentioned: mentionGate.effectiveWasMentioned,
-    shouldBypassMention,
+    effectiveWasMentioned: decision.effectiveWasMentioned,
+    shouldBypassMention: decision.shouldBypassMention,
   };
 }
